Update docs-elastic-dev-publish.yml to use project ID fallback
We are having an issue where the build fail. due to updating the builders. I originally tried storing the new secret in this repo but it appears that the actions will not take a combination of passed-in and local secrets. 

This PR attempts to circumvent the issue in the interim, when the migration to the new names is taking place by having the original value as a fallback.