Update docs-elastic-dev-publish.yml
This PR changes the fallback order in place for the `vercel-project-id`. We've had issues where changing the variables that are sent from the consuming actions trigger a second build that fails after merging. My assumption is that because the `VERCEL_PROJECT_ID_DOCS_DEV` is shared across all repos that the fallback is not working properly. 

Will test and revert if any new issues arise.